fix: fixed error response

+++ API/Errors/ApiException.cs
@@ -1,8 +1,10 @@
+using System.Net;
+
 namespace API.Errors
 {
     public class ApiException : ApiResponse
     {
-        public ApiException(int statusCode, string message = null, string details = null) : base(statusCode, message)
+        public ApiException(int statusCode, string message = null, string details = null) : base((int)statusCode, message)
         {
             Details = details;
         }
